requirements.txt
openai>=1.3.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
